feat: Went through solid tutorial and updated some things
There were a couple of places where I'd missed cleanup or was doing
something that wasn't as performant as it could have been. Hopefully now
it is all very pretty and lovely. Who knows :P seemed to work ahahha.

Also fixed width on the textarea... I spelt it wrong... the humiliation
xD