feat: Move Timer into its own component
There are 3 obvious sections to the current page and before I start
adding anymore xD I needed a tidy up. It was getting a little busy for
my brain ahah.

Anyway created a Countdown context which was pretty fun and very similar
to react. Intuitive to use and feels tidy tbh. Happy with it so far.