feat: Add cat theme.
There is no theme changer and I am.... lets just say very unsure on the
colours HOWEVER. The background image is perfection and I will hear
no slander.

Stickers and the prize image now change based on the theme. This brings
me joy.

I am... lets say not happy with how I did the styling themeing on the
stickers. But I don't think its terrible so might just have to move on.

ANYWAY. Glory be to cats. Next is the minimalist version.