feat: Add minimalist theme
Again not 100% happy with this. Do need to remember to change the timer
audio alarm on these. Loud violins doesn't feel very minimalist xD Maybe
it is! Who knows. Anyway for me its a little much ahahha.

Had quite some fun poking the theme to get this to handle fonts
seperately which honestly is very useful.